services/anemia/requirements.txt
tensorflow==2.13.0
numpy==1.24.3
matplotlib==3.7.2
opencv-python>=4.8.0
pillow>=10.0.0
scikit-learn==1.3.0
pandas==2.0.3
mediapipe==0.10.7
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.2
